cram: Bring reader and writer to htslib parity for CRAM 2.0–4.0
Add full read/write support for CRAM versions 2.0, 2.1, 3.0, 3.1, and
4.0, reaching feature parity with htslib/samtools. This includes:

Reader changes:
- Legacy codec support (beta, gamma, Golomb, Huffman, subexp) for v2.x
- CRAM 4.0 EOF detection, slice header encoding, and codec offsets
- Async reader parity across all versions
- Overflow protection for VLQ integer readers
- Fix ref_seq_id encoding and codec argument signedness

Writer changes:
- Version-gated CRC32 checksums and qs_seq_orient encoding
- fqzcomp HAVE_QMAP, HAVE_QTAB, MULTI_PARAM encoder features
- Owned tag data and reference-free mode support
- Name tokenizer tok_dup and fqzcomp DO_REV support

API additions:
- Reader::query_unmapped() and Reader::for_each_record()
- IndexedReader delegation of the above methods

Test infrastructure:
- Programmatic in-memory test data creation (no on-disk test files)
- Integration tests for round-trip, reader methods, query APIs, and
  writer options across all CRAM versions

+++ noodles-cram/src/io/writer/record.rs
@@ -26,7 +26,7 @@ pub struct Record {
     pub(crate) mate_flags: MateFlags,
     pub(crate) mate_reference_sequence_id: Option<usize>,
     pub(crate) mate_alignment_start: Option<Position>,
-    pub(crate) template_length: i32,
+    pub(crate) template_length: i64,
     pub(crate) mate_distance: Option<usize>,
     pub(crate) data: Vec<(Tag, Value)>,
     pub(crate) features: Vec<Feature>,
